// src/app/models/product.ts
export interface Product {
      product_name?: string;
      product_category_id?: string;
      product_id_sku?: string;
      product_price?: number;
      product_short_desc?: string;
      product_description?: string;
      supplier_id?: string;
      product_quantity?: number;
      product_status?: boolean;
      product_image?: any;
      product_id?: string;
      product_thumb_image?: any;
}